validators: move validation functions to own file
